Again Merge/Sync omnielement/firefly to master (#36)
* Add old icon

* Add logo for potential wiki use or anything

* 优化一些贴图

* 贴图优化

* 魔茧顶点光照

* 还是贴图优化

* Water optimization

* BGM实装

* hotfix

* Try MothEye Equip Stat Effects

When MothEye is equipped and the player is in the firefly biome, holding the ScaleWingBlade will gain a buff.
Trying the getter/setter method I learned in Java
~Setnour6

Co-Authored-By: Omnielement <103014909+Omnielement@users.noreply.github.com>

* Try full code use instead of getter/setter

It worked better on testing. I was hoping that reducing the amount of code for each item would work, but it lead to some problems. Maybe that can be tried again in the future.
~Setnour6

Co-Authored-By: Omnielement <103014909+Omnielement@users.noreply.github.com>

* ...and add note

* 完善青缎矿系列装备。还有天穹树一，二层的一些物块表现

* Debug, 优化效果

* Enhance YggdrasilTownBGM

Now the song sounds less generically structured.
What's a good name for the song? Maybe... "Town in the Tree"? "Great Tree Town"? I'll think about it.

* Reflect changes from MothEye-Continue-Rework

* Fix recipe result + knockback balancing

* Set rarity to White, from orange, given item stats

Co-authored-by: Omnielement <877183629@qq.com>
Co-authored-by: yiyang233 <1787058089@qq.com>
Co-authored-by: Ye_you <49426220+dfgasdfg@users.noreply.github.com>
Co-authored-by: Omnielement <103014909+Omnielement@users.noreply.github.com>
Co-authored-by: 28271655 <28271655@qq.com>