CMoth Value Increase + MEye bonuses are hardmode only
This is so the weapons can be used normally against the wall of flesh, then receive the extra power in the firefly biome. This also encourages the Corrupted Moth to be fought again (if fought already in pre-hardmode) for the extra weapon benefits, as long as you are in the firefly biome. I even see story and lore potential.